good enough ™ (#1363)
